modify eth_feeHistory to take []u64 instead of []f64 to be compatible with FEVM RPC

+++ ethers-core/src/utils/mod.rs
@@ -82,7 +82,7 @@ pub const WEI_IN_ETHER: U256 = U256([0x0de0b6b3a7640000, 0x0, 0x0, 0x0]);
 /// The number of blocks from the past for which the fee rewards are fetched for fee estimation.
 pub const EIP1559_FEE_ESTIMATION_PAST_BLOCKS: u64 = 10;
 /// The default percentile of gas premiums that are fetched for fee estimation.
-pub const EIP1559_FEE_ESTIMATION_REWARD_PERCENTILE: f64 = 5.0;
+pub const EIP1559_FEE_ESTIMATION_REWARD_PERCENTILE: u64 = 5;
 /// The default max priority fee per gas, used in case the base fee is within a threshold.
 pub const EIP1559_FEE_ESTIMATION_DEFAULT_PRIORITY_FEE: u64 = 3_000_000_000;
 /// The threshold for base fee below which we use the default priority fee, and beyond which we
